Exam 1
This exam was used to make sure we understood all the basics of C++ like basic loops and data structures. This lab was used to read a file of information on DVD, Music, and TV. This added them to lists and then printed out the mean, median, and mode on a separate file.